API integration done. Data displayed successfully. (Image Load will be added)

+++ app/src/main/java/com/example/spacexfan/model/modelproperties/Height.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
    
 data class Height (
 
-   @SerializedName("meters") var meters : Int
+   @SerializedName("meters") var meters : Double
 
 )
